Clean up document.py
Move things not of general interest into other modules. Add more
docstrings.
