+++ bokeh/core/has_props.py
@@ -1,4 +1,5 @@
-'''
+''' Provide a base class for objects that can have declarative, typed,
+serializable properties.
 
 .. note::
     These classes form part of the very low-level machinery that implements
@@ -35,7 +36,7 @@ if IPython:
     class _BokehPrettyPrinter(RepresentationPrinter):
         def __init__(self, output, verbose=False, max_width=79, newline='\n'):
             super(_BokehPrettyPrinter, self).__init__(output, verbose, max_width, newline)
-            self.type_pprinters[HasProps] = lambda obj, p, cycle: obj._bokeh_repr_pretty_(p, cycle)
+            self.type_pprinters[HasProps] = lambda obj, p, cycle: obj._repr_pretty(p, cycle)
 
 _ABSTRACT_ADMONITION = '''
     .. note::
@@ -288,6 +289,7 @@ class HasProps(with_metaclass(MetaHasProps, object)):
             True, if properties are structurally equal, otherwise False
 
         '''
+
         # NOTE: don't try to use this to implement __eq__. Because then
         # you will be tempted to implement __hash__, which would interfere
         # with mutability of models. However, not implementing __hash__
@@ -710,7 +712,7 @@ class HasProps(with_metaclass(MetaHasProps, object)):
         '''
         return self.__class__(**self._property_values)
 
-    def _bokeh_repr_pretty_(self, p, cycle):
+    def _repr_pretty(self, p, cycle):
         '''
 
         '''
