+++ bokeh/sphinxext/bokeh_model.py
@@ -37,7 +37,7 @@ from docutils.parsers.rst.directives import unchanged
 
 from sphinx.errors import SphinxError
 
-from ..model import Viewable
+from ..model import MetaModel
 from .bokeh_directive import BokehDirective, py_sig_re
 from .templates import MODEL_DETAIL
 
@@ -70,8 +70,8 @@ class BokehModelDirective(BokehDirective):
         if model is None:
             raise SphinxError("Unable to generate reference docs for %s, no model '%s' in %s" % (model_name, model_name, module_name))
 
-        if type(model) != Viewable:
-            raise SphinxError("Unable to generate reference docs for %s, model '%s' is not a subclass of Viewable" % (model_name, model_name))
+        if type(model) != MetaModel:
+            raise SphinxError("Unable to generate reference docs for %s, model '%s' is not of type MetaModel" % (model_name, model_name))
 
         model_obj = model()
