+++ sphinx/source/docs/reference/model.rst
@@ -3,5 +3,15 @@
 bokeh.model
 ===========
 
+.. use autodata below for model_class_reverse_map otherwise it prints the
+.. huge actual reverse map contents in the rendered docs
+
 .. automodule:: bokeh.model
   :members:
+
+    .. autoclass: MetaModel
+
+        .. autodata:: model_class_reverse_map
+            :annotation: = {}
+
+            Stores a mapping from __view_model__ names to Bokeh model classes
